api/serializers.py
from rest_framework import serializers
from .validators import checkStrongPassword
from .models import Task, User

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model=Task
        fields="__all__"

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"