requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
Django==2.2
django-cors-headers==3.7.0
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==19.7.1
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==3.3.1
